Login and Registration system complete !!!!!!!!!!

+++ Assets/Scripts/Base Game/ScoreManager.cs
@@ -7,7 +7,7 @@ public class ScoreManager : MonoBehaviour
 {
     public TextMeshProUGUI scoreText;
 
-    private int score = 0;
+    private long score = 0;
 
     public void UpdateScore(int newScore)
     {
